backend/requirements.txt
asn1crypto==1.5.1
async-timeout==5.0.1
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.2.1
cryptography==38.0.4
Flask==2.3.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
freezegun==1.5.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mirakuru==2.6.1
packaging==25.0
pg8000==1.31.4
pluggy==1.6.0
port-for==0.7.4
psutil==7.0.0
psycopg==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
pytest-pgsql==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
redis==7.0.1
scramp==1.4.6
six==1.17.0
SQLAlchemy==2.0.43
testing.common.database==2.0.3
testing.postgresql==1.3.0
typing_extensions==4.15.0
ulid-py==1.1.0
Werkzeug==3.1.3
